\name{Choices}
\alias{Choices}
\alias{Choices-class}
\title{The Choices class}
\description{
  The Choices class contains a vector of Gambles class
  objects.
}
\note{
  Choices contains a vector of Gambles class objects. Each
  vector element stores a decision making choice.
}
\section{Slots}{
  \describe{ \item{\code{choices}:}{Object of class
  \code{"vector"}, containing Gambles objects for decision
  makers to choose from.} }
}

