\name{compound_invariance_pwf}
\alias{compound_invariance_pwf}
\title{The compound invariance probability weighting function.}
\usage{
compound_invariance_pwf(par, x)
}
\arguments{
  \item{par}{vector, a vector containing parameters for the
  pwf}

  \item{x}{numeric, the x value}
}
\description{
Compound invariance probability weighting function.
}
\references{
Prelec, D. (1998). The probability weighting function.
Econometrica, 60(3), 497-528.
}

