\name{constant_relative_sensitivity_pwf}
\alias{constant_relative_sensitivity_pwf}
\title{The constant relative sensitivity probability weighting function.}
\usage{
constant_relative_sensitivity_pwf(par, x)
}
\arguments{
  \item{par}{vector, parameters for the probability
  weighting function.}

  \item{x}{numeric, the x value}
}
\description{
Constant relative sensitivity probability weighting
function.
}
\references{
Abdellaoui, M., L'Haridon, O., & Zank, H. (2010).
Separating curvature and elevation: A parametric
probability weighting function. Journal of Risk and
Uncertainty, 41(1), 39-65.
}

