% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{energy_group_map}
\alias{energy_group_map}
\title{Panos-tuotostoimialojen vastaavuus energiatilaston toimialaryhmiin}
\usage{
energy_group_map()
}
\value{
data.frame: `industry` (panos-tuotostaulukon koodi) ja `group`
  (taulukon `tene/11wy` toimialaluokka).
}
\description{
Panos-tuotostoimialojen vastaavuus energiatilaston toimialaryhmiin
}
