\name{eta.bd}
\alias{eta.bd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Predictions of Slope functions eta as detailed in
as detailed in Monajemi and Donoho 2018 (Sparsity/Undersampling Tradeoffs 
in Anisotropic Undersampling)
}
\description{
Given the undersampling delta, a field, ambient dimenion of one block
$N$, and number of blocks B, this function return the sparsity
at the compressed sensing Phase transition for block-diagonal (anisotropic) 
undersampling
}
\usage{
eta.bd <- function(deltas,field)
%- maybe also 'usage' for other objects documented here.
\arguments{
  
  \item{delta}{
     the undersampling ratio: Number of data points/number of unknowns
    }
  \item{field}{
    the coefficient field. "R", "Pos", "Bnd", "Q", "O" .
    }

}

\details{
}

\value{
The returned value is list
result <- list(eta=f.deltas,alpha=alpha,beta=beta, slope=s);

}
\references{
Sparsity/Undersampling Tradeoff in anisotropic undersampling, with applications in MR imaging/
spectroscopy.
https://academic.oup.com/imaiai/advance-article/doi/10.1093/imaiai/iay013/5183349
}
\author{
Hatef Monajemi (monajemi@stanford.edu) and 
David Donoho   (donoho@stanford.edu)
}
\note{
This code comes with absolutely no gaurantee. The user accepts to be 
responsible for the results generated by this code.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
eta <- eta.bd(1/2,"Pos");
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~CompressedSensing ~Ell1 ~AnisotropicUndersampling ~ BlockDiagonal}
