\name{exponential_power_pwf}
\alias{exponential_power_pwf}
\title{The exponential power probability weighting function.}
\usage{
exponential_power_pwf(par, x)
}
\arguments{
  \item{par}{vector, a vector containing parameters for the
  pwf}

  \item{x}{numeric, the x value}
}
\description{
Exponential-power probability weighting function.
}
\references{
Stott, H. P. (2006). Cumulative prospect theory's
functional menagerie. Journal of Risk and Uncertainty,
32(2), 101-130.
}

