\name{Gamble}
\alias{Gamble}
\alias{Gamble-class}
\title{The Gamble class}
\description{
  The Gamble class
}
\note{
  The Gamble class represents a gamble, i.e. a choice
  option containing various outcomes whose probabilities
  need to sum to unity.
}
\section{Slots}{
  \describe{ \item{\code{gamble_id}:}{Object of class
  \code{"numeric"}, containing the unique id of the
  gamble.} \item{\code{outcomes}:}{Object of class
  \code{"vector"}, containing the possible outcomes of the
  gamble.} }
}

