\name{hyperbolic_logarithm_pwf}
\alias{hyperbolic_logarithm_pwf}
\title{The hyperbolic logarithm probability weighting function.}
\usage{
hyperbolic_logarithm_pwf(par, x)
}
\arguments{
  \item{par}{vector, a vector containing parameters for the
  pwf}

  \item{x}{numeric, the x value}
}
\description{
Hyperbolic-logarithm probability weighting function.
}
\references{
Stott, H. P. (2006). Cumulative prospect theory's
functional menagerie. Journal of Risk and Uncertainty,
32(2), 101-130.
}

