% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfin.R
\name{jsonstat_to_df}
\alias{jsonstat_to_df}
\title{Muunna json-stat2-vastaus siistiksi data.frameksi}
\usage{
jsonstat_to_df(js)
}
\arguments{
\item{js}{`jsonlite::fromJSON(..., simplifyVector = FALSE)`-muodossa
luettu json-stat2-objekti.}
}
\value{
data.frame, jossa ulottuvuuksien koodit, selitteet ja `value`.
}
\description{
Muunna json-stat2-vastaus siistiksi data.frameksi
}
