\name{kt_pwf}
\alias{kt_pwf}
\title{The Tversky and Kahneman (1992) probability weighting function.}
\usage{
kt_pwf(par, x)
}
\arguments{
  \item{par}{vector, a vector containing parameters for the
  pwf}

  \item{x}{numeric, the x value}
}
\description{
Tversky and Kahneman's (1992) probability weighting
function.
}
\references{
Tversky, A., & Kahneman, D. (1992). Advances in prospect
theory: Cumulative representation of uncertainty. Journal
of Risk and Uncertainty, 5(4), 297-323.
}

