\name{linear_in_log_odds_pwf}
\alias{linear_in_log_odds_pwf}
\title{The linear in log odds probability weighting function.}
\usage{
linear_in_log_odds_pwf(par, x)
}
\arguments{
  \item{par}{vector, a vector containing parameters for the
  pwf}

  \item{x}{numeric, the x value}
}
\description{
Linear in log odds probability weighting function.
}
\references{
Wu, G., & Gonzalez, R. (1996). Curvature of the probability
weighting function. Management Science, 42(12), 1676-1690.
}

