\name{neo_additive_pwf}
\alias{neo_additive_pwf}
\title{neo_additive_pwf}
\usage{
  neo_additive_pwf(parameters, x)
}
\arguments{
  \item{parameters}{vector, a vector containing parameters
  for the pwf}

  \item{x}{numeric, the x value}
}
\description{
  Neo-additive probability weighting function.
}
\references{
  Stott, H. P. (2006). Cumulative prospect theory's
  functional menagerie. Journal of Risk and Uncertainty,
  32(2), 101-130.
}

