\name{Probability_Weighting}
\alias{Probability_Weighting}
\alias{Probability_Weighting-class}
\title{The Probability_Weighting class}
\description{
  The Probability_Weighting class
}
\note{
  Probability_Weighting stores both the functional form and
  parameter specification for a probability weighting
  function.
}
\section{Slots}{
  \describe{ \item{\code{probability_function}:}{Object of
  class \code{"text"}, containing a text string that
  specifies the functional form of the probability
  weighting function.} \item{\code{parameters}:}{Object of
  class \code{"vector"}, containing numeric values for the
  parameter specifications associated with the probability
  weighting function.} }
}

