\docType{class}
\name{ProbWeight-class}
\alias{ProbWeight-class}
\title{The ProbWeight class.}
\description{
The ProbWeight class stores both the form and parameter
specification for a probability weighting function.
}
\note{
A function (also called ProbWeight) has been defined to
create an instance of this class.
}
\section{Slots}{
  \describe{ \item{\code{fun}:}{Object of class
  \code{"text"}, containing a text string that specifies
  the functional form of the probability weighting
  function.} \item{\code{par}:}{Object of class
  \code{"vector"}, containing the parameter specifications
  for the probability weighting function.} }
}
\seealso{
{\code{\link{ProbWeight}}}
}

