% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pt_coefficients}
\alias{pt_coefficients}
\title{Teknologiakertoimet}
\usage{
pt_coefficients(Z, x)
}
\arguments{
\item{Z}{Välituotekäyttömatriisi (toimittaja x käyttäjä).}

\item{x}{Tuotosvektori.}
}
\value{
`A = Z diag(x)^-1`; nollatuotoksen sarakkeet nollia.
}
\description{
Teknologiakertoimet
}
