% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pt_ghosh}
\alias{pt_ghosh}
\title{Ghoshin käänteismatriisi (tarjontapuolen malli)}
\usage{
pt_ghosh(Z, x)
}
\arguments{
\item{Z}{Välituotekäyttömatriisi.}

\item{x}{Tuotosvektori.}
}
\value{
`(I - B)^-1`, missä `B = diag(x)^-1 Z` (allokaatiokertoimet).
}
\description{
Ghoshin käänteismatriisi (tarjontapuolen malli)
}
