% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pt_impact}
\alias{pt_impact}
\title{Kysyntävaikutukset}
\usage{
pt_impact(sys, f, type2 = FALSE, ...)
}
\arguments{
\item{sys}{`pt_system`.}

\item{f}{Nimetty vektori (toimialakoodi -> milj. e) tai täysi vektori.}

\item{type2}{Käytetäänkö suljettua mallia (kotitalouksien kulutuksen
kerrannaisvaikutukset mukana).}

\item{...}{Argumentit [pt_closed_system()]:lle.}
}
\value{
Lista: `summary` (kokonaisvaikutukset) ja `industries`
  (toimialoittaiset vaikutukset).
}
\description{
Laskee loppukysynnän muutoksen `f` (milj. e perushintaan, kotimaiset
tuotteet) vaikutukset tuotokseen, arvonlisäykseen, palkkoihin, työllisiin
ja tuontiin.
}
