% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{pt_industry_group}
\alias{pt_industry_group}
\title{Toimialat ja niiden ryhmät}
\usage{
pt_industry_group(codes)
}
\arguments{
\item{codes}{Toimialakoodit (esim. `"C17"`, `"D35"`).}
}
\value{
Merkkijonovektori: `"Alkutuotanto"`, `"Kaivostoiminta"`,
  `"Teollisuus"`, `"Energia- ja vesihuolto"`, `"Rakentaminen"`, `"Palvelut"`.
}
\description{
Toimialat ja niiden ryhmät
}
