% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{pt_is_industry}
\alias{pt_is_industry}
\title{Onko toimiala teollisuutta (TOL C)}
\usage{
pt_is_industry(codes, mining = FALSE, energy = FALSE)
}
\arguments{
\item{codes}{Toimialakoodit.}

\item{mining}{Otetaanko mukaan kaivostoiminta (B).}

\item{energy}{Otetaanko mukaan energiahuolto (D35).}
}
\description{
Onko toimiala teollisuutta (TOL C)
}
