% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pt_leontief}
\alias{pt_leontief}
\title{Leontiefin käänteismatriisi}
\usage{
pt_leontief(A)
}
\arguments{
\item{A}{Teknologiakerroinmatriisi.}
}
\value{
`(I - A)^-1`.
}
\description{
Leontiefin käänteismatriisi
}
