% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pt_linkages}
\alias{pt_linkages}
\title{Rasmussenin kytkennät ja Ghoshin eteenpäinkytkentä}
\usage{
pt_linkages(sys)
}
\arguments{
\item{sys}{`pt_system`.}
}
\value{
data.frame: taaksepäin- ja eteenpäinkytkentäindeksit (keskiarvo 1).
  Toimiala on avaintoimiala, jos molemmat ovat yli 1.
}
\description{
Rasmussenin kytkennät ja Ghoshin eteenpäinkytkentä
}
