% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pt_multipliers}
\alias{pt_multipliers}
\title{Kertoimet: tuotos, arvonlisäys, työllisyys ja tuonti}
\usage{
pt_multipliers(sys)
}
\arguments{
\item{sys}{`pt_system`.}
}
\value{
data.frame toimialoittain.
}
\description{
Tyypin I kertoimet kertovat, paljonko yhden euron (miljoonan euron)
loppukysyntä toimialan tuotteisiin synnyttää koko taloudessa tuotosta,
arvonlisäystä, työllisiä ja välituotetuontia (suorat + välilliset vaikutukset).
}
