% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pt_price_channels}
\alias{pt_price_channels}
\title{Hintavaikutuksen välittymiskanavat}
\usage{
pt_price_channels(sys, industry, domestic = numeric(), imported = numeric())
}
\arguments{
\item{sys}{`pt_system`.}

\item{industry}{Toimialakoodi, jonka hintavaikutus puretaan.}

\item{domestic}{Nimetty vektori kotimaisten toimialojen hinnanmuutoksista
(osuutena, 0.1 = +10 \%), esim. `c(D35 = 0.1)`.}

\item{imported}{Nimetty vektori tuontituotteiden hinnanmuutoksista
tuotekoodeittain (14yp), esim. `c("35" = 0.1, "19" = 0.1)`.}
}
\value{
data.frame: `channel`, `label`, `contribution`, `share`
  (osuus kokonaisvaikutuksesta), suurimmasta pienimpään.
}
\description{
Purkaa [pt_price_model()]:n kokonaisvaikutuksen toimialalle `industry`
niihin toimialoihin, joiden kustannusten kautta vaikutus välittyy:
`total_j = sum_k d_k * L_N[k, j]`, missä `d` on suora vaikutus ja
`L_N = (I - A[N,N])^-1`. Toimialan oma rivi sisältää suoran vaikutuksen.
}
