% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{pt_tables}
\alias{pt_tables}
\title{Luettelo StatFinin panos-tuotostaulukoista}
\usage{
pt_tables()
}
\value{
data.frame: taulukon tunnus, nimi ja päivitysaika.
}
\description{
Luettelo StatFinin panos-tuotostaulukoista
}
