% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{pt_years}
\alias{pt_years}
\title{Saatavilla olevat panos-tuotosvuodet}
\usage{
pt_years()
}
\value{
Kokonaislukuvektori vuosista, joilta symmetrinen
  panos-tuotostaulukko (14yn) on saatavilla.
}
\description{
Saatavilla olevat panos-tuotosvuodet
}
