% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfin.R
\name{statfin_meta}
\alias{statfin_meta}
\title{Hae StatFin-taulukon metatiedot}
\usage{
statfin_meta(table, cache = TRUE)
}
\arguments{
\item{table}{Taulukon polku StatFin-tietokannassa, esim. `"pt/14yn.px"`.
Hakemistopolku (esim. `"pt/"`) palauttaa hakemiston taulukkoluettelon.}

\item{cache}{Käytetäänkö välimuistia.}
}
\value{
Lista, jossa `title` ja `variables` (koodit, arvot ja selitteet).
}
\description{
Hae StatFin-taulukon metatiedot
}
