% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfin.R
\name{statfin_variables}
\alias{statfin_variables}
\title{Muuttujien arvot ja selitteet taulukkomuodossa}
\usage{
statfin_variables(table, cache = TRUE)
}
\arguments{
\item{table}{Taulukon polku StatFin-tietokannassa, esim. `"pt/14yn.px"`.
Hakemistopolku (esim. `"pt/"`) palauttaa hakemiston taulukkoluettelon.}

\item{cache}{Käytetäänkö välimuistia.}
}
\value{
data.frame: `variable`, `variable_text`, `code`, `text`.
}
\description{
Muuttujien arvot ja selitteet taulukkomuodossa
}
