\name{Utility}
\alias{Utility}
\alias{Utility-class}
\title{The Utility class}
\description{
  The Utility class
}
\note{
  Utility stores both the functional form and parameter
  specification for a utility function.
}
\section{Slots}{
  \describe{ \item{\code{utility_function}:}{Object of
  class \code{"text"}, containing a text string that
  specifies the functional form of the utility function.}
  \item{\code{parameters}:}{Object of class
  \code{"vector"}, containing numeric values for the
  parameter specifications associated with the utility
  function.} }
}

