\name{compound_invariance_pwf}
\alias{compound_invariance_pwf}
\title{compound_invariance_pwf}
\usage{
  compound_invariance_pwf(parameters, x)
}
\arguments{
  \item{parameters}{vector, a vector containing parameters
  for the pwf}

  \item{x}{numeric, the x value}
}
\description{
  Compound invariance probability weighting function.
}
\references{
  Prelec, D. (1998). The probability weighting function.
  Econometrica, 60(3), 497-528.
}

