\name{hyperbolic_logarithm_pwf}
\alias{hyperbolic_logarithm_pwf}
\title{hyperbolic_logarithm_pwf}
\usage{
  hyperbolic_logarithm_pwf(parameters, x)
}
\arguments{
  \item{parameters}{vector, a vector containing parameters
  for the pwf}

  \item{x}{numeric, the x value}
}
\description{
  Hyperbolic-logarithm probability weighting function.
}
\references{
  Stott, H. P. (2006). Cumulative prospect theory's
  functional menagerie. Journal of Risk and Uncertainty,
  32(2), 101-130.
}

