\name{kt_pwf}
\alias{kt_pwf}
\title{kt_pwf}
\usage{
  kt_pwf(parameters, x)
}
\arguments{
  \item{parameters}{vector, a vector containing parameters
  for the pwf}

  \item{x}{numeric, the x value}
}
\description{
  Tversky and Kahneman's (1992) probability weighting
  function.
}
\references{
  Tversky, A., & Kahneman, D. (1992). Advances in prospect
  theory: Cumulative representation of uncertainty. Journal
  of Risk and Uncertainty, 5(4), 297-323.
}

