\name{neo_additive_pwf}
\alias{neo_additive_pwf}
\title{The neo-additive probability weighting function.}
\usage{
neo_additive_pwf(par, x)
}
\arguments{
  \item{par}{vector, a vector containing parameters for the
  pwf}

  \item{x}{numeric, the x value}
}
\description{
Neo-additive probability weighting function.
}
\references{
Stott, H. P. (2006). Cumulative prospect theory's
functional menagerie. Journal of Risk and Uncertainty,
32(2), 101-130.
}

