\docType{class}
\name{Utility-class}
\alias{Utility-class}
\title{The Utility class.}
\description{
The Utility class stores both the functional form and
parameter specifications for a utility function.
}
\note{
A wrapper function (also called Utility) can be used to
create an instance of this class.
}
\section{Slots}{
  \describe{ \item{\code{fun}:}{Object of class
  \code{"text"}, containing a text string that specifies
  the functional form of the utility function.}
  \item{\code{par}:}{Object of class \code{"vector"},
  containing the parameter specifications for the utility
  function.} }
}
\seealso{
{\code{\link{Utility}}}
}

